\name{recode}

\alias{recode_rf}


\title{
Recode SNP data 
}

\description{
Recoding ...
}


\usage{
recode_rf(d, biallelic=TRUE, missflags="-")
}
 
\arguments{
\item{d}{data.frame with combined ref and fld data created with \code{\link{combine_rf}}}
\item{biallelic}{logical}
\item{missflag}{character}
}

\value{
data.frame
}


